requirements.txt
tenacity
fastapi
uvicorn
openai
python-dotenv
gspread
google-auth
oauth2client
google-auth-oauthlib
google-oauth2-tool
requests
pydantic
supabase
python-jose[cryptography]
pytest-asyncio
python-pptx
whisper
pathlib
aiofiles
